Change VAT rate type from int to float across the system
remp/crm#3277

+++ src/Models/PaymentItem/PostalFeePaymentItem.php
@@ -28,7 +28,7 @@ final class PostalFeePaymentItem implements PaymentItemInterface
 
     public function __construct(
         private readonly ActiveRow $postalFee,
-        int $vat,
+        float $vat,
         int $count = 1,
         array $meta = [],
     ) {
